Make project screen topbar resemble mockups
