+++ ui/DesignSystem/Primitive/Icon/Issue.svelte
@@ -1,0 +1,19 @@
+<script>
+  export let style = null;
+</script>
+
+<svg
+  {style}
+  width="24"
+  height="24"
+  viewBox="0 0 24 24"
+  fill="var(--color-gray)"
+  xmlns="http://www.w3.org/2000/svg">
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M16 7H8C7.44772 7 7 7.44772 7 8V16C7 16.5523 7.44772 17 8 17H16C16.5523
+    17 17 16.5523 17 16V8C17 7.44772 16.5523 7 16 7ZM8 5C6.34315 5 5 6.34315 5
+    8V16C5 17.6569 6.34315 19 8 19H16C17.6569 19 19 17.6569 19 16V8C19 6.34315
+    17.6569 5 16 5H8ZM11 8H13V13H11V8ZM13 14H11V16H13V14Z" />
+</svg>
