--- ui/Screen/Project/Feed.svelte
@@ -1,7 +1,0 @@
-<script>
-  import { Flex, Title } from "../../DesignSystem/Primitive";
-</script>
-
-<Flex align="left">
-  <Title variant="big">Feeds</Title>
-</Flex>
